# api/openapi/http-api.yaml
openapi: "3.0.3"
info:
  title: Balda GameServer
  description: |
    Balda GameServer API methods and models

    ### Headers
    | Header | Description |
    |--------|-------------|
    | **X-API-Key** | API key is a special token that the client needs to provide when making API calls. |
    | **X-API-User** | User's ID |
    | **X-API-Session** | Session ID |
  version: 1.0.0

servers:
  - url: http://127.0.0.1:9666/balda/api/v1

paths:
  /signup:
    post:
      operationId: signup
      summary: Sign-up request
      tags:
        - Signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "200":
          description: Response for signup request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignupResponse"
        "400":
          description: Error when signup
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth:
    post:
      operationId: auth
      summary: Auth request
      tags:
        - Auth
      security:
        - APIKeyHeader: []
        - APIKeyQueryParam: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "200":
          description: Response for auth request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Error when auth
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /session/ping:
    post:
      operationId: ping
      tags: [Session]
      summary: Keepalive ping
      description: |
        Mirrors cm_ping (id=1). POST not GET — mutates session TTL.
        Returns 204 with X-Server-Time header instead of a JSON body
        to minimize bandwidth on frequent pings (every ping_delay ms).
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
      responses:
        '204':
          description: Session alive.
          headers:
            X-Server-Time:
              $ref: '#/components/headers/X-Server-Time'
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /player/state/{uid}:
    get:
      operationId: getPlayerStateUID
      summary: Get user state
      parameters:
        - name: uid
          in: path
          description: Player ID
          required: true
          schema:
            type: string
            format: uuid
            description: The unique identifier of the player
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerState"
        "400":
          description: Error when get player state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
    APIKeyQueryParam:
      type: apiKey
      in: query
      name: api_key

  responses:
    Unauthorized:
      description: Missing or invalid session (mirrors err_code -100).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: -100
            message: "Session required"

  schemas:
    Error:
      type: object
      properties:
        code:
          description: Error code
          type: integer
        message:
          description: Human-readable error message
          type: string

    ErrorResponse:
      type: object
      properties:
        status:
          description: the HTTP response code
          example: 401
          type: integer
        message:
          description: a human-readable explanation of the error
          example: Bad Authentication data.
          type: string
        type:
          description: Type of the error message
          example: OAuthException
          type: string

    SignupRequest:
      type: object
      required:
        - firstname
        - lastname
        - email
        - password
      properties:
        firstname:
          description: User's first name
          example: John
          type: string
        lastname:
          description: User's last name
          example: Smith
          type: string
        email:
          description: User's email
          example: js@example.org
          type: string
          format: email
        password:
          description: User's password
          example: some_password
          type: string

    SignupResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/Player"

    Player:
      type: object
      properties:
        uid:
          description: Player's ID in the system
          type: string
          format: uuid
        firstname:
          description: Player's first name
          example: John
          type: string
        lastname:
          description: Player's last name
          example: Smith
          type: string
        sid:
          description: Player's session ID
          example: 6c160004-35c8-45b1-a4da-0ab67076afde
          type: string
        key:
          description: Player's API Key that the client needs to provide when making API calls.
          example: some_api_key
          type: string

    AuthRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          description: User's email
          example: js@example.org
          type: string
        password:
          description: User's password
          example: some_password
          type: string

    AuthResponse:
      type: object
      properties:
        player:
          $ref: "#/components/schemas/Player"

    PlayerState:
      type: object
      properties:
        uid:
          description: Player's ID in the system
          type: string
          format: uuid
        nickname:
          description: Generated nickname in the system
          example: JohnSmith
          type: string
        exp:
          description: Player's exp points in the system
          example: 100
          type: integer
          format: int64
        lives:
          description: Player's lives count in the system
          example: 3
          type: integer
          format: int64
        flags:
          description: Some Player's flags.
          example: 10
          type: integer
          format: int64
        game_id:
          description: Game's ID (if game_id empty or absent then player in the lobby)
          example: 144
          type: string
          format: uuid
          
  headers:
    X-Request-ID:
      description: Monotonically increasing client sequence number. Mirrors command_num. Server echoes it back.
      schema:
        type: integer
        minimum: 1
        format: int64
    X-Server-Time:
      description: Server Unix timestamp in milliseconds.
      schema:
        type: integer
        format: int64
    Retry-After:
      description: Seconds until the rate-limited or cooled-down action may be retried.
      schema:
        type: integer
    ETag:
      description: Entity tag for conditional GET caching.
      schema:
        type: string
    Cache-Control:
      description: Cache directive.
      schema:
        type: string
    X-RateLimit-Limit:
      description: Maximum requests allowed in the current window.
      schema:
        type: integer
    X-RateLimit-Remaining:
      description: Requests remaining in the current window.
      schema:
        type: integer
    X-RateLimit-Reset:
      description: When the current rate-limit window resets (ISO 8601).
      schema:
        type: string
        format: date-time

  parameters:
    RequestIdHeader:
      in: header
      name: X-Request-ID
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1